+++ packages/utils/crates/utils-misc/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod async_iterable;
 pub mod utils;
+pub mod time;
 
 #[cfg(feature = "wasm")]
 pub mod wasm {